feat(3f): extract debug windows to DevToolsUI class
